+++ papyrus/basepage.py
@@ -1,13 +1,13 @@
 import pathlib
 import datetime
 import os
-import re
-
-import frontmatter
 
 class BasePage():
-    def __init__(self, input_path):
+    def __init__(self, input_path, meta, content):
         self.input_path = input_path
+        self.meta = meta
+        self.content = content
+
         self.output_path = None
         self.layout = None
         self.url = None
@@ -22,47 +22,25 @@ class BasePage():
         self.excerpt_html = None
 
     def build(self, pandoc, config):
-        with self.input_path.open() as infile:
-            self.content_text = infile.read()
-        # parse and remove front matter
-        front_meta, content = frontmatter.parse(self.content_text)
-        # parse and remove back matter
-        back_meta, content = self.parse_back_matter(content)
-        meta = front_meta | back_meta
-        self.title, content = self.extract_title(content)
+        self.title, content = self.extract_title(self.content)
         self.excerpt_text, self.content_text = \
             self.get_excerpt_content(content, config['EXCERPT_SEPARATOR'])
-        self.layout = meta['layout'] if 'layout' in meta else 'post'
-        self.date = self.get_date(meta, self.input_path)
-        self.url = self.get_url(meta)
-        self.props = meta['props'].split(" ") if "props" in meta else []
-        self.tags = ('tags' in meta and meta['tags'].split(" ")) or []
+
+        self.layout = self.meta['layout'] if 'layout' in self.meta else 'post'
+        self.date = self.get_date(self.meta, self.input_path)
+        self.url = self.get_url(self.meta)
+        self.props = self.meta['props'].split(" ") if "props" in self.meta else []
         self.output_path = self.get_output_path(config['DIR_PUBLISH'])
-
-        if not self.is_current():
-            self.generate_html(pandoc)
-        self.wordcount = pandoc.countwords(self.input_path)
-
-    @staticmethod
-    def parse_back_matter(content):
-        back_meta = {}
-        ii = []
-        for mo in re.finditer(r"^---\s*$", content, re.MULTILINE):
-            ii.append(mo.start())
-
-        if (len(ii)) == 2 and len(content) == ii[1] + 3:
-            back_matter = content[ii[0]:]
-            back_meta, _ = frontmatter.parse(back_matter)
-            content = content[:ii[0]].strip()
-
-        return back_meta, content
 
     @staticmethod
     def extract_title(content):
         ii_first_newline = content.find("\n")
-        first_line = content[0:ii_first_newline]
-        content = content[ii_first_newline:].strip()
-        return first_line.replace("#", "").strip(), content
+        first_line = content[0:ii_first_newline].strip()
+        if first_line.startswith("#"):
+            content = content[ii_first_newline:].strip()
+            return first_line.replace("#", "").strip(), content
+
+        return "", content
 
     def is_public(self):
         if not self.props:
@@ -114,7 +92,7 @@ class BasePage():
         splittext = full_text.split(excerpt_sep, maxsplit=1)
 
         if len(splittext) < 2:
-            return "", splittext[0].strip()
+            return None, splittext[0].strip()
 
         return splittext[0].strip(), splittext[0] + splittext[1].strip()
 
@@ -131,12 +109,12 @@ class BasePage():
         else:
             return True
 
+    def render(self, jinja_env, site_meta, page_meta):
+        raise NotImplementedError()
+
     def write(self, jinja_env, site_meta):
         if not self.is_current():
-            print("Writing:", self)
-            template = jinja_env.get_template("{}.html".format(self.layout))
-            page = self.get_page_metadata()
-            template_out = template.render(content=self.content_html, page=page, post=self, site=site_meta)
+            template_out = self.render(jinja_env, site_meta, self.get_page_metadata())
             self.output_path.parent.mkdir(parents=True, exist_ok=True)
             self.output_path.write_text(template_out)